[MRG] modify names in examples, run_examples and most workflows (workflow f…

+++ macapype/nodes/extract_brain.py
@@ -298,7 +298,7 @@ class AtlasBREXInputSpec(CommandLineInputSpec):
         "a,0,0", usedefault=True, desc="msk", position=7, argstr="-msk %s",
         mandatory=True)
 
-    dil = traits.Float(desc="dil", argstr="-dil %f")
+    dil = traits.Int(desc="dil", argstr="-dil %d")
 
     nrm = traits.Enum(1, 2, desc="nrm", argstr="-nrm %d")
